+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Change Log:
 
-# 0.0.1 Setting up package
 
+### 0.0.2 
+* Combine meshes
+* UVs for cubes
+### 0.0.1 
+* Setting up package
+* Basic importer